requirements.txt
jsonschema==4.21.1
ruamel.yaml==0.18.6
requests==2.31.0
